% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{add_media}
\alias{add_media}
\title{Add a media line to a campaign}
\usage{
add_media(schedule, campaign, media)
}
\arguments{
\item{schedule}{A schedule object}

\item{campaign}{Campaign index number}

\item{media}{A media object}
}
\value{
A schedule containing the new media line
}
\description{
Add a media line to a campaign
}
