\docType{methods}
\name{addPar}
\alias{addPar}
\title{Add real parameters to the instance}
\arguments{
  \item{object}{An optimInstance object}

  \item{par}{A parameter in the optimSMS object of the
  instance}

  \item{indices}{The values of the indices for the value}

  \item{values}{The numeric value for the parameter}
}
\description{
  Add real parameters to the instance
}

