% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_curves.R
\name{calc_curve_parameters}
\alias{calc_curve_parameters}
\title{Estimates diminishing returns parameters for an entire set of channel level curves}
\usage{
calc_curve_parameters(channel_curves)
}
\arguments{
\item{channel_curves}{a tibble made with make_channel_curves()}
}
\description{
Estimates diminishing returns parameters for an entire set of channel level curves
}
