% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{campaign}
\alias{campaign}
\title{Defines a new campaign object}
\usage{
campaign(name = character(), product = character())
}
\arguments{
\item{product}{}
}
\description{
Defines a new campaign object
}
