% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{create_schedule}
\alias{create_schedule}
\title{Creates a new schedule object}
\usage{
create_schedule(name, start_date, end_date, granularity = "month")
}
\arguments{
\item{granularity}{}
}
\description{
Creates a new schedule object
}
