\docType{data}
\name{dsFields}
\alias{dsFields}
\title{Lists of fields for each data.frame in the smsOptim object}
\format{List of 8
 $ fItem   : chr [1:5] "id" "symbol" "tag" "sDes" ...
 $ fConst  : chr [1:2] "value" "aux"
 $ fSet    : chr [1:6] "loc" "inSet" "setType" "dataCom" ...
 $ fMeasure: chr [1:5] "nature" "units" "dataType" "setInd" ...
 $ fVar    : chr [1:5] "ind" "integer" "positive" "binary" ...
 $ fPar    : chr [1:3] "group" "ind" "dataCom"
 $ fEq     : chr [1:6] "nature" "relation" "domain" "group" ...
 $ fTerm   : chr [1:11] "eq" "side" "parent" "nature" ...}
\usage{
  dsFields
}
\description{
  Lists of fields for each data.frame in the smsOptim
  object
}
\keyword{datasets}

