\docType{methods}
\name{exportPars}
\alias{exportPars}
\alias{exportPars,exportPars-method}
\title{Export instance parameters}
\arguments{
  \item{.Object}{}

  \item{format}{}

  \item{file}{}
}
\description{
  Export instance parameters to other files for optimisers
  or data analysis
}
\details{
  Accepted formats: gdx (GAMS), xlsx, xml
}
\author{
  Emilio L. Cano
}

