\docType{methods}
\name{SMSsets}
\alias{instanceId}
\alias{instanceId-method,}
\alias{InstancePar-method}
\alias{InstancePars}
\alias{InstanceSet}
\alias{InstanceSet-method}
\alias{InstanceVar-method}
\alias{InstanceVars}
\alias{SMSconsts}
\alias{SMSeqs}
\alias{SMSeqs-method}
\alias{SMSpars}
\alias{SMSpars-method}
\alias{SMSsets}
\alias{SMSsets-method}
\alias{SMSvars}
\alias{SMSvars-method}
\title{Get SMS sets}
\arguments{
  \item{object}{An smsOptim object}

  \item{object}{An smsOptim object}

  \item{object}{An smsOptim object}

  \item{object}{An smsOptim object}

  \item{object}{An smsOptim object}

  \item{.Object}{optimInstance object}

  \item{object}{An optimInstance object}

  \item{set}{A set symbol in the SMS}

  \item{object}{An optimInstance object}

  \item{par}{A set symbol in the SMS}

  \item{object}{An optimInstance object}

  \item{par}{A set symbol in the SMS}
}
\value{
  list with the elements id, name, sDes, lDes
}
\description{
  Gets the data.frame with the sets of the SMS

  Gets the data.frame with the constants of the SMS

  Gets the data.frame with the variables of the SMS

  Gets the data.frame with the parameters of the SMS

  Gets the data.frame with the equations of the SMS

  Gets the id, name, short description and long description
  of an instance

  Gets a vector with the elements of a set in the instance

  Gets a data.frame with the parameter values and indices

  Gets a data.frame with the variable values and indices
}

