% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_curves.R
\name{get_curve_points}
\alias{get_curve_points}
\title{Reads an entire directory of Media Mix Navigator zip files}
\usage{
get_curve_points(path)
}
\arguments{
\item{path}{directory path (must only contain Media Mix Navigator zip files)}
}
\value{
a tibble of budget scenarios
}
\description{
Reads an entire directory of Media Mix Navigator zip files
}
