% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_curves.R
\name{get_dim_rets_params}
\alias{get_dim_rets_params}
\title{Estimates parameters for negative exponential diminishing returns curves}
\usage{
get_dim_rets_params(curve, start_a = 5000, start_b = 50000)
}
\arguments{
\item{start_b}{}
}
\description{
Estimates parameters for negative exponential diminishing returns curves
}
