\name{getAliases}
\alias{getAliases}
\title{Get model aliases}
\arguments{
  \item{object}{An optimSMS object}

  \item{format}{The output format}

  \item{compact}{If true, no long descriptions or data
  comments are included}
}
\description{
  Get the set aliases of a model
}
\details{
  For human readable formats, e.g., tex, the order is
  alphabetical. For machine readable formats, e.g., gams,
  the order is the one of the data.frame.
}

