\docType{methods}
\name{getConsts}
\alias{getConsts}
\alias{getConsts,optimSMS-method}
\title{Get constants in the model, such as scalars}
\arguments{
  \item{object}{}

  \item{format}{}

  \item{compact}{}
}
\value{
  Character string with the formatted constants
}
\description{
  Get constants in the model, such as scalars
}
\author{
  emilio
}

