\name{getEntityDomain}
\alias{getEntityDomain}
\title{Get the expression of the domain for an entity}
\usage{
  getEntityDomain(.Object, entity, entId, format = "tex")
}
\arguments{
  \item{.Object}{}

  \item{entity}{}

  \item{entId}{}

  \item{format}{}
}
\value{
  A character with the expression of the domain
}
\description{
  A parameter or variable can be defined over specific
  subsets related to their indices.
}
\author{
  Emilio L. Cano
}

