\name{getEq}
\alias{getEq}
\title{Get the expression of an equation}
\arguments{
  \item{object}{an optimSMS object}

  \item{getid}{Identification of the item}

  \item{format}{Output format of the equation}

  \item{only}{Allows to get only a right hand side (rExpr)
  or left hand side (lExpr) of an equation}
}
\description{
  Get the expression of an equation
}

