\name{getEqs}
\alias{getEqs}
\title{Get formatted equations}
\arguments{
  \item{object}{An optimSMS object}

  \item{format}{The output format}

  \item{compact}{If \code{TRUE} descriptions are not
  provided}

  \item{eqEnv}{Equation environment for LaTeX format}
}
\description{
  Get all the equations of a SMS in a given format
}

