\name{getPrimeSuperset}
\alias{getPrimeSuperset}
\title{Get the symbol of the superset when the symbol is aux (prime)}
\usage{
  getPrimeSuperset(symbol)
}
\arguments{
  \item{symbol}{character string with a symbol}

  \item{format}{An accepted format}
}
\description{
  Returns the appropriate set symbol
}

