\name{getSymbol}
\alias{getSymbol}
\alias{getSymbol,}
\alias{getSymbol-method}
\title{Get the symbol of an item}
\arguments{
  \item{object}{an optimSMS object}

  \item{entity}{Variable, parameter, set, constant}

  \item{getid}{Identification of the item}
}
\description{
  Get the symbol of an item
}

