\name{hasBrothers}
\alias{hasBrothers}
\title{Find out if an equation term has child terms}
\arguments{
  \item{object}{An smsOptim object}

  \item{geteq}{An equation id}

  \item{getid}{A term id in the equation}

  \item{age}{}
}
\description{
  Age expects '>' or '<' to get posterior and prior
  brothers respectively
}

