\docType{methods}
\name{importPar}
\alias{importPar}
\title{Add real parameters to the instance}
\arguments{
  \item{object}{An optimInstance object}

  \item{par}{A parameter in the optimSMS object of the
  instance}

  \item{values}{A data frame with the parameter values and
  their indices in the instance}
}
\description{
  Add a whole data.frame object with the values of the
  parameter and their indices.
}

