\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize-method}
\title{Initializes an optimSMS class object}
\description{
  Initializes an optimSMS class object

  Initializes an optimInstance class object
}
\examples{
mySMS <- new("optimSMS", "example SMS",
		"This is a short description for example SMS.",
		"This is a LONG description for exampleSMS.")
mySMS <- new("optimSMS", "example SMS",
		"This is a short description for example SMS.",
		"This is a LONG description for exampleSMS.")
myInstance <- new("optimInstance", mySMS)
}

