\docType{methods}
\name{initialize}
\alias{initialize}
\title{Initializes an optimInstance class object}
\description{
  Initializes an optimInstance class object
}
\examples{
mySMS <- new("optimSMS", "example SMS",
		"This is a short description for example SMS.",
		"This is a LONG description for exampleSMS.")
myInstance <- new("optimInstance", mySMS)
}

