\name{is.mdset}
\alias{is.mdset}
\title{Check if a set is a multidimensional set}
\usage{
  is.mdset(instance, set)
}
\arguments{
  \item{instance}{An instance objecty}

  \item{set}{A character with a set symbol}
}
\value{
  TRUE if set is a multidimensional set
}
\description{
  Check if a set is a multidimensional set
}
\author{
  emilio
}

