\name{is.subset}
\alias{is.subset}
\title{Check if a set is a subet}
\usage{
  is.subset(instance, set)
}
\arguments{
  \item{instance}{An instance objecty}

  \item{set}{A character with a set symbol}
}
\value{
  TRUE if set is a subset
}
\description{
  Check if a set is a subet
}
\author{
  emilio
}

