% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_curves.R
\name{make_channel_curves}
\alias{make_channel_curves}
\title{Tranforms Media Mix Navigator data into channel-level response curves}
\usage{
make_channel_curves(curve_points)
}
\arguments{
\item{curve_points}{a tibble made with get_curve_points()}
}
\description{
Tranforms Media Mix Navigator data into channel-level response curves
}
