% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_scraper.R
\name{make_scenarios}
\alias{make_scenarios}
\title{Builds a set of budget scenarios to extract from Media Mix Navigator}
\usage{
make_scenarios(
  min_budget = 1,
  max_budget = 10,
  step_budget = 1,
  category_option = 1,
  appeal_option = 1,
  percent_sales_online_option = 1,
  brand_size = 100,
  risk_option = 1
)
}
\arguments{
\item{min_budget}{minimum budget}

\item{max_budget}{maximum budget}

\item{step_budget}{step size between min_budget and max_budget}

\item{category_option}{Item number in the dropdown list (All Categories is #1)}

\item{appeal_option}{Item number in the dropdown list}

\item{percent_sales_online_option}{Item number in the dropdown list}

\item{brand_size}{brand size in £m}

\item{risk_option}{Item number in the dropdown list}
}
\value{
a tibble of budget scenarios
}
\description{
Builds a set of budget scenarios to extract from Media Mix Navigator
}
