% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{media}
\alias{media}
\title{Defines a new media object}
\usage{
media(
  name = character(),
  type = character(),
  weight_metric = "GRP",
  alpha = numeric(),
  beta = numeric(),
  decay = numeric()
)
}
\arguments{
\item{decay}{}
}
\description{
Defines a new media object
}
