\name{newInstance}
\alias{newInstance}
\title{Instance creation}
\usage{
  newInstance(sms)
}
\arguments{
  \item{sms}{an optimSMS object with the model.}
}
\value{
  An empty optimInstance object, ready to add items
}
\description{
  Function to create an empty optimInstance object, ready
  to add items
}
\author{
  Emilio L. Cano
}

