\name{newSMS}
\alias{newSMS}
\title{Symbolic Model Specification (SMS) creation}
\usage{
  newSMS(name, sDes = "", lDes = "")
}
\arguments{
  \item{name}{A name for the model.}

  \item{sDes}{A short description for the model}

  \item{lDes}{A long description for the model}
}
\value{
  An empty optimSMS object, ready to add items
}
\description{
  Function to create an empty optimSMS object, ready to add
  items
}
\author{
  Emilio L. Cano
}

