\name{optimInstance}
\alias{optimInstance}
\title{optimInstance Class}
\description{
  instance of a Symbolic Model Specification.
}
\details{
  This class contains the related Symbolic Model
  Specification (optimSMS object). Besides, the
  implementation contains the variables and equations
  implemented, the actual sets in the model, and the values
  for the parameters.
}

