% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{optimise_schedule}
\alias{optimise_schedule}
\title{Optimises a schedule}
\usage{
optimise_schedule(schedule, max_budget = NULL, step = NULL)
}
\arguments{
\item{step}{}
}
\value{
An optimised schedule
}
\description{
Optimises a schedule
}
