% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_curves.R
\name{read_curve_point}
\alias{read_curve_point}
\title{Reads curve information from a single zip downloaded from Media Mix Navigator}
\usage{
read_curve_point(file)
}
\arguments{
\item{file}{zip filename}
}
\description{
Reads curve information from a single zip downloaded from Media Mix Navigator
}
