% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{schedule_demo}
\alias{schedule_demo}
\title{Creates a demo schedule for testing purposes}
\usage{
schedule_demo()
}
\value{
A populated schedule object
}
\description{
Creates a demo schedule for testing purposes
}
\examples{
test_schedule <- schedule_demo()
}
