% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_mix_navigator_scraper.R
\name{scrape_navigator}
\alias{scrape_navigator}
\title{Downloads a scenario from Media Mix Navigator}
\usage{
scrape_navigator(
  category_option = NULL,
  appeal_option = NULL,
  percent_sales_online_option = NULL,
  brand_size = NULL,
  budget = NULL,
  risk_option = NULL,
  remDr = NULL
)
}
\arguments{
\item{category_option}{Item number in the dropdown list (All Categories is #1)}

\item{appeal_option}{Item number in the dropdown list}

\item{percent_sales_online_option}{Item number in the dropdown list}

\item{brand_size}{Brand size in £m}

\item{budget}{Budget in £m}

\item{risk_option}{Item number in the dropdown list}

\item{remDr}{A selenium connection, usually made with connect_selenium()}
}
\value{
NULL (a zip file will save to the download directory)
}
\description{
Note: Exclude 'Please Choose...' from the item count
}
