% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{set_media_laydown}
\alias{set_media_laydown}
\title{Sets spend or weight for a media line or a single date within a media line and recalculates dependent metrics}
\usage{
set_media_laydown(
  schedule = NULL,
  campaign = NULL,
  media = NULL,
  values = NULL,
  date_index = NULL,
  metric = "spend"
)
}
\arguments{
\item{schedule}{A schedule object}

\item{campaign}{Campaign index number}

\item{media}{Media index number}

\item{values}{Values to set for spend or weight}

\item{metric}{"spend" or "weight"}

\item{index}{Optional date index to use if setting only one value}
}
\value{
An amended schedule object
}
\description{
Sets spend or weight for a media line or a single date within a media line and recalculates dependent metrics
}
