\docType{methods}
\name{show}
\alias{show}
\alias{show.optimInstance-method}
\alias{show.optimSMS-method}
\title{Shows the optimSMS object}
\arguments{
  \item{object}{optimSMS object to show}

  \item{object}{optimInstance object to show}
}
\description{
  Shows the optimSMS object

  Shows the optimInstance object
}
\author{
  Emilio L. Cano

  Emilio L. Cano
}

