% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{sum_uplift_schedule}
\alias{sum_uplift_schedule}
\title{Returns the total uplift generated by a schedule}
\usage{
sum_uplift_schedule(schedule)
}
\arguments{
\item{schedule}{}
}
\value{
An uplift number
}
\description{
Returns the total uplift generated by a schedule
}
