\docType{data}
\name{symTrans}
\alias{symTrans}
\title{Dataframe containing translations for the different formats}
\format{'data.frame':	10 obs. of  4 variables:
 $ sms : chr  "eq" "lte" "gte" "lt" ...
 $ gams: chr  "=e=" "=l=" "=g=" "" ...
 $ tex : chr  " = " " \\leq " " \\geq " "<" ...
 $ expr: chr  "=" "<=" ">=" "<" ...}
\usage{
  symTrans
}
\description{
  Dataframe containing translations for the different
  formats
}
\keyword{datasets}

