% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{test_increment}
\alias{test_increment}
\title{Increments spend for one media on one date and returns the uplift difference generated. Used internally for optimisation runs.}
\usage{
test_increment(schedule, campaign, media, date_index, increment)
}
\arguments{
\item{increment}{}
}
\description{
Increments spend for one media on one date and returns the uplift difference generated. Used internally for optimisation runs.
}
