% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{unnest_schedule}
\alias{unnest_schedule}
\title{Unnests a schedule object into a tibble where each row is one media on one date. Used internally by optimR for optimisation runs.}
\usage{
unnest_schedule(schedule)
}
\arguments{
\item{schedule}{}
}
\description{
Unnests a schedule object into a tibble where each row is one media on one date. Used internally by optimR for optimisation runs.
}
