\name{wProblem}
\alias{wProblem}
\title{Generic method for writing the file for the solver or interface}
\arguments{
  \item{object}{an optimInstance object}

  \item{filename}{a string with the filename}

  \item{format}{an admitted format to write}

  \item{solver}{The solver to use}

  \item{parsFile}{File containing the parameter values}

  \item{eqsFile}{File containing the equations}
}
\description{
  Generic method for writing the file for the solver or
  interface
}

