% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_s3.R
\name{unnest_schedule}
\alias{unnest_schedule}
\title{Unnests a schedule object into a tibble. Used internally by optimR for optimisation runs and charting.}
\usage{
unnest_schedule(schedule, level = "date")
}
\arguments{
\item{schedule}{An optimR media schedule}

\item{level}{The level to unnest to. Can be "date" (fully unnested), or "media", which will leave dates nested}
}
\description{
Unnests a schedule object into a tibble. Used internally by optimR for optimisation runs and charting.
}
